.github/scripts/requirements.txt
openapi-schema-validator>=0.6.3
pyyaml>=6.0.2